Height Checker
Students are asked to stand in non-decreasing order of heights for an
annual photo. Return the minimum number of students that must move in
order for all students to be standing in non-decreasing order of height.
Notice that when a group of students is selected they can reorder in any
possible way between themselves and the non selected students remain on
their seats.

Example:
Input: heights = [5,1,2,3,4]
Output: 5

Input: heights = [1,2,3,4,5]
Output: 0