Validate Binary Search Tree (74/75)
Given a binary tree, determine if it is a valid binary search tree (BST)

Assume a BST is defined as follows:

The left subtree of a node contains only nodes with keys less than the
node's key. The right subtree of a node contains only nodes with keys
greater than the node's key. Both the left and right subtrees must also
be binary search trees.

Example:
    2
   / \
  1   3

Input: [2,1,3]
Output: true