Kth Largest Element in a Stream (10/10)
Design a class to find the kth largest element in a stream. Note that
it is the kth largest element in the sorted order, not the kth
distinct element.
Your KthLargest class will have a constructor which accepts an integer
k and an integer array nums, which contains initial elements from the
stream. For each call to the method KthLargest.add, return the element
representing the kth largest element in the stream.

NOTE: I resolved the algorith using an ArrayList, I have to improve
      it usign a binary tree.